requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
faiss-cpu
